[icu][assembly] Fixes up the intl assembly
It turns out that some products have an existing
configuration that is not expressible in the current
assembly config for the `intl` subsystem.

This change handles that issue by introducing
the missing configuration option. The actual
solution would probably be to reduce the config
variation, but that's a more extensive change
that I'd like to keep separate.

Bug: 117974, b/296136021
Change-Id: If480f7579d582f9bf2af9112f60ea5b4af0ff0db
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/902176
Reviewed-by: Aidan Wolter <awolter@google.com>
Commit-Queue: Filip Filmar <fmil@google.com>