[logs] Fix symbolizer
This patch fixes two bugs in the symbolizer:
* Newlines are present for empty symbolize output
* --raw wasn't working because it wasn't passed
from ffx to log-command.
This patch also gets rid of the ..Default::default()
that was being used to catch default variables
being passed at compile-time so this won't result in
runtime failures in the future.

Fixed: 132260
Change-Id: Id19e70703c9b7632e70b7a1e5482afac1a19ab2c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/902642
Reviewed-by: Christopher Johnson <crjohns@google.com>
Reviewed-by: Miguel Flores <miguelfrde@google.com>
Commit-Queue: Brian Bosak <bbosak@google.com>