[fdf][examles] Remove demo driver base test
This test is way too simple to add any value. The driver was not doing
anything except a log message on start. The test was not validating
anything. The demo number driver that exists here has a lot more value
and there is nothing here that the demo number one doesn't have. So we
can remove this example/test.

Bug: 132131
Change-Id: I3cc767ad80211c0cb34652c4cc381abe019a7741
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/903400
Reviewed-by: Sarah Chan <spqchan@google.com>
Commit-Queue: Novin Changizi <novinc@google.com>