[diagnostics] Drain the socket when not measuring
The benchmark is sufficiently CPU intensive that other threads can
impact the benchmarks. Prior to this CL, a thread is spawned to drain
the socket to ensure it doesn't overflow, but this thread could affect
timings. (See https://fxrev.dev/901921 for an example of a change which
one would expect to not have an impact on the benchmark, but it actually
causes a regression.)

The change here makes sure draining of the socket happens outside of the
measurement.

Fixed: 131474
Change-Id: I3e27886960f319a1153cc11a84b2fbdb320aca9a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/902618
Commit-Queue: Chris Suter <csuter@google.com>
Reviewed-by: Miguel Flores <miguelfrde@google.com>