[profiler][fidl] Add Targeting Components And Tests
This adds support to the profiler FIDL protocol for specifying component
and test targets.

Profiling a component will cause each task it launches to be profiled.
Profiling a test will launch a test via test manager and attach to it
via its job.

The actual implementation to do so is stubbed out for now and will
follow in coming patches.

This adds ComponentConfig and TestConfig as a type of TargetConfig. It
also pushes the vector into specifically launching adhoc Tasks as
attempting to profile multiple components or tests at once is somewhat
complicated and there isn't an immediate use case for it.

Change-Id: I3ac20023776d024ed1a74b105ad6df0db7bca5d7
Bug: b/293459772
Bug: b/294081668
Bug: b/284190829
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/872191
Reviewed-by: Christopher Johnson <crjohns@google.com>
Reviewed-by: Brian Hamrick <bhamrick@google.com>
API-Review: Christopher Johnson <crjohns@google.com>
Commit-Queue: Gwen Mittertreiner <gmtr@google.com>