[bedrock] Support standalone namespace entries in serve_processargs
This serves the needs of "/pkg" directories and directory/storage
capabilities and the like. You can add an Open capability (or something
that is TryIntoOpen) at a path, and the Open capability will be opened
when someone opens that path.

A subtle requirement here is that we need to delay exercising the open
capability until the user writes a message onto the namespace entry, as
seen in `add_directory_helper` in component manager. That has been
replicated here and in fact doesn't affect regular namespaced objects.

Change-Id: I777cc813f90e75d76890cfaac0f8556ba5914f2b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/900279
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Yifei Teng <yifeit@google.com>
Reviewed-by: Yegor Pomortsev <ypomortsev@google.com>