[heapdump][ffx] Print a usage error without the "internal error" preamble
This error is not a bug, so the "BUG: An internal command error
occurred" preamble is confusing to the user. By using ffx_bail, such a
preamble is not printed.

In addition, this makes this error consistent with the other user-facing
errors generated by this source file.

Change-Id: I8fbe4a49def7c7ed5c5927a954caf8af01290a70
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/903259
Reviewed-by: Étienne J. Membrives <etiennej@google.com>
Commit-Queue: Fabio D'Urso <fdurso@google.com>