[icu_data] Remove unneeded target
This target was only used at one place, and was
a forwarding target anyways. So we can remove it.

I noticed this late in working on
https://fuchsia-review.git.corp.google.com/c/fuchsia/+/903498,
and I didn't want to destroy the presubmit results to
add this inconsequential change.

Bug: 117974
Change-Id: I65b27f713107eac3b467b878abfea5c52681b9de
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/904098
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Filip Filmar <fmil@google.com>
Reviewed-by: Mukesh Agrawal <quiche@google.com>