[netstack3] Use a single ID for ICMP sockets.
Note that this change does not guard against bugs if you use a wrong
type of socket ID. The rationale is that this is a transitory state
and the refactor to use `DatagramSocketSpec` to implement ICMP sockets
will change most of the existing functions. There is currently also no
usage of ICMP sockets elsewhere, so keeping the throwaway work minimal
as long as the tests are passing.

Bug: 125482
Change-Id: I365156f82f5d589537b2ab64f826cdac2b61e2cb
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/902099
Fuchsia-Auto-Submit: Zeling Feng <zeling@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Peter Johnston <peterjohnston@google.com>