[netstack3] Add UNREACHABLE neighbor state
Add the UNREACHABLE state for dynamic neighbor entries, which is
conceptually similar to the PROBE state: packets continued to be sent to
the cached link address, and neighbor probes are actively sent to
resolve the neighbor. However, in UNREACHABLE, the probes are multicast
rather than unicast, to allow for discovering a link-layer address
change for the neighbor.

Probes are transmitted as long as packets continue to be sent to the
neighbor, but with exponential backoff rather than on a fixed retransmit
timer. When packets are not sent, the netstack also stops sending
probes.

Bug: 35185
Test: netstack3-core-test
Change-Id: Ie91609ba62b344592a499c69c20ae27363c85383
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/898678
Reviewed-by: Ghanan Gowripalan <ghanan@google.com>
Commit-Queue: Peter Johnston <peterjohnston@google.com>
Fuchsia-Auto-Submit: Peter Johnston <peterjohnston@google.com>
Reviewed-by: Zeling Feng <zeling@google.com>