[rust-3p] Update to 5.2.0
This updates the notify crate to 5.2.0. Unfortunately while it looks
like they attempted to include the license file, it appears that it
doesn't actually include the license text, so we'll need to maintain our
own fork for now.

Along the way, we also need to update our windows-sys to 0.45.0. However
we still have a few things still using windows-sys 0.42.0, so it keeps
an empty version of that crate as well until we finish migrating off
of it.

Change-Id: I928596635b4ff643d6f49fec0ca27a6217b3b00c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/904140
Reviewed-by: Adam Perry <adamperry@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Erick Tryzelaar <etryzelaar@google.com>
Reviewed-by: Randall Bosetti <rlb@google.com>
Reviewed-by: Erick Tryzelaar <etryzelaar@google.com>