[ffx] Add debugging for ssh connection failure
To understand fxb/127900 better, we need some debugging on the
initial ssh connection.  Depending on what information comes out
of this, we can improve the ssh handling (including a more
refined approach to timing out on the parsing the ssh output).

This also adds an ssh timeout so we get a useful report if we hang
indefinitely.

Bug: 288918487
Change-Id: I429c1cdc9fe76837671dd82d9166c42593687f00
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/903500
Reviewed-by: Casey Dahlin <sadmac@google.com>
Reviewed-by: Clayton Wilkinson <wilkinsonclay@google.com>
Commit-Queue: Steven Grady <slgrady@google.com>