[netstack3] Ensure SyncCtx outlives NonSyncCtx
If NonSyncCtx outlives SyncCtx, we can encounter a panic due to
referency IDs (while SyncCtx is being torn down, it expects to
hold the sole copy of referency IDs). This is a particularly thorny
footgun in tests.

To avoid this, we avoid impling Clone on NonSyncCtx, and instead expose
accessor methods for the SyncCtx and NonSyncCtx on the bindings Ctx
struct, which itself is Clone.

Test: netstack3-tests
Change-Id: I0cbb7f1f9a02d2284ed39d0d0e51f318ba63760d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/903940
Commit-Queue: Evan Wildenhain <wildenhain@google.com>
Reviewed-by: Bruno Dal Bo <brunodalbo@google.com>
Fuchsia-Auto-Submit: Evan Wildenhain <wildenhain@google.com>