[assembly] add a FileRelativePathBuf crate
This crate allows the deserialization of paths that are relative
to the file they're given in, and then resolving those paths
against the path to that containing file.

It also allows the conversion of paths into a file-relative form,
given that the path to the file and the path to the containing
file are both given from the same base.

Change-Id: Ie781413aebe8bc52f08fa5b99035636ac3692c63
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/903503
Reviewed-by: Aidan Wolter <awolter@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Aaron Wood <aaronwood@google.com>