[zircon][qemu] fix serial port when using the -g option to fx qemu
A previous refactor had broken the automatic redirection of serial
to the console when using the -g switch, which enables a graphics
window.

Generally refactor the code dealing with the -serial switch to be a bit
more robust. Notably --no-serial overrides everything, and -S defaults
to 'stdio'.

Change-Id: I26b56c1eece027a6839eac77f59e4d8a49fdaf9a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/904400
Reviewed-by: Joshua Seaton <joshuaseaton@google.com>
Reviewed-by: Dwayne Slater <dwayneslater@google.com>
Commit-Queue: Travis Geiselbrecht <travisg@google.com>