[assembly] Remove vestiges of legacy intl configuration
Required firming up ICU config as well, to emphasize that
there are products that are not interested in the intl
config. Mostly this is the "basic" products.

Furthermore, added the defaults for each of intl, setui
and icu to be used if no special values are needed.

Bug: 117974
Change-Id: I189ceb204989a639263772b4b1a7e0dd85666df8
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/905711
Reviewed-by: Aaron Wood <aaronwood@google.com>
Commit-Queue: Filip Filmar <fmil@google.com>
Fuchsia-Auto-Submit: Filip Filmar <fmil@google.com>