[netstack3] Add dual stack conn state
Encode whether a connected socket is connected in `ThisStack` or the
`OtherStack` using an enum. This provides additional type safety when
performing dual stack operations because it allows us to match on the
`DualStackConnState` and explicitly handle the dual-stack case.

As a fallout from this CL, we now can clearly see the remaining TODO's
associated with dual-stack datagram sockets.

Fixed: 132333

Change-Id: I130fd9822e1ed99be6feeb1282b5be01dc403c7b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/903098
Reviewed-by: Bruno Dal Bo <brunodalbo@google.com>
Reviewed-by: Zeling Feng <zeling@google.com>
Commit-Queue: Jeff Martin <martinjeffrey@google.com>