[components] Don't set default path in testing decl builders
Capability decls builders used in tests were setting arbitrary
default source_paths that are not the documented default
of "/svc/{name}" for protocols and services.

Setting a default value in the decl makes the false assumption that
the test will add it that to the outgoing dir, which is not true
for tests that don't use RoutingTest.

For protocols, the path "/svc/foo" is added by default to the outgoing
directory only in RoutingTest, defined in CM. However, the decl builders
are defined in the cm_rust library and can be used in any other test.
A generic library should not depend on details of a specific test.

Change-Id: I30dc73ebdf9ac9caf98d54281acc509bcfee2efe
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/904364
Reviewed-by: Yifei Teng <yifeit@google.com>
API-Review: Yegor Pomortsev <ypomortsev@google.com>
Commit-Queue: Yegor Pomortsev <ypomortsev@google.com>