[fdf] Composite node spec rebind on DFv2
Forward rebind requests to the CompositeNodeManager
if the node is a composite.

Implement DriverRunner::RequestRebindFromDriverIndex()
so that the CompositeNodeSpecManager can send rebind
requests to the Driver Index.

Implement CompositeNodeSpecV2::RemoveImpl() so that
the composite node and parent set collector can be removed
for rebinding.

Tests: driver-runner-tests
Bug: 124976
Change-Id: I6d5e736c936f8484c850f573247b610ccacb663b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/901466
Commit-Queue: Sarah Chan <spqchan@google.com>
Reviewed-by: Novin Changizi <novinc@google.com>