[assembly] Ramdisk mode is enabled via CLI arg
Instead of placing the ramdisk mode in the product config, we now pass
it to product assembly as a CLI argument. This will allow us to use the
same product config between products even if they opt to put the
packages in a ramdisk in the ZBI (such as netboot).

Bug: b/297090553
Change-Id: I3d37da798397c88e6dad7560f648bbe232e6c005
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/903905
Reviewed-by: Aaron Wood <aaronwood@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Aidan Wolter <awolter@google.com>