[elfldltl] elfldltl::SegmentWithVmo::GetMutableMemory
This is an object to wire up elfldltl::LoadInfoMutableMemory using
elfldltl::SegmentWithVmo so that a new copy-on-write VMO child is
created for each mutable segment as needed.  This returns a Memory
object using a modified elfldltl::MappedVmoFile to map in each new
VMO so relocations can be applied to its contents.

Change-Id: Id4cb42f162319acc4571e1010f1230900fc0adeb
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/900480
Commit-Queue: Roland McGrath <mcgrathr@google.com>
Reviewed-by: Caslyn Tonelli <caslyn@google.com>
Fuchsia-Auto-Submit: Roland McGrath <mcgrathr@google.com>