[logs] Delete log_ng
This is now redundant and can be removed. The code in this plugin
was copied to the previous "ffx log" plugin due to FIDL APIs not being in the SDK and eventually will be converted away from a dual-mode plugin
when it is possible to do so and all FIDL dependencies have been
exported to the SDK and versioned.

Bug: 121413
Change-Id: I86aae74650556709b61dde91ad106a0fa94ef451
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/907323
Reviewed-by: Casey Dahlin <sadmac@google.com>
Reviewed-by: Christopher Johnson <crjohns@google.com>
Commit-Queue: Brian Bosak <bbosak@google.com>
Reviewed-by: Miguel Flores <miguelfrde@google.com>