[fxtest][rewrite] Implement --parallel
This implements the --parallel flag in a way that is not compatible with
the original fx test. Originally it would control the per-suite parallel
flag, but now it controls the number of simultaneous suites run on a
single target.

The ffx test --parallel flag overrides the suite's settings, but it is
most commonly used to ensure that suites that have internal crosstalk do
not fail to run. We should always respect the tests.json parallel flag,
which is what this CL does for the new fx test.

Fixed: b/295340779
Change-Id: I1b0b1c3ff328d24a22b73bf97fc054672eccddf6
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/904204
Reviewed-by: Kevin Cho <chok@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Christopher Johnson <crjohns@google.com>