[fuchsia-controller] Improve union support.
This reworks unions to support root-level unions from Fuchsia Controller calls
to FIDL methods, but also to handle unions in a way that is clearer than the
initial implementation, so what goes into the encoding end is the same
as what comes out of the decoding end.

See the fidl server tests for details on what this looks like
(particularly the file server implementation).

Bugs: b/297275231, b/296640503

Change-Id: I924c6364bbe107665b9eaad4174167fa0bb4262b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/906932
Reviewed-by: Mitchell Kember <mkember@google.com>
Commit-Queue: Andrew Davies <awdavies@google.com>